requirements.txt
fastapi==0.95.0
uvicorn==0.17.6
sqlalchemy==1.4.50
psycopg2-binary==2.9.10
pydantic==1.9.1
python-dotenv==1.0.0
faker==20.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1 